move overlap filter to feature subfolder
