+++ src/openms/include/OpenMS/PROCESSING/SPECTRAMERGING/SpectraMerger.h
@@ -15,7 +15,7 @@
 #include <OpenMS/CONCEPT/LogStream.h>
 #include <OpenMS/CONCEPT/ProgressLogger.h>
 #include <OpenMS/DATASTRUCTURES/DefaultParamHandler.h>
-#include <OpenMS/PROCESSING/DATAREDUCTION/SplineInterpolatedPeaks.h>
+#include <OpenMS/PROCESSING/MISC/SplineInterpolatedPeaks.h>
 #include <OpenMS/KERNEL/BaseFeature.h>
 
 #include <vector>
