+++ src/openms/includes.cmake
@@ -44,7 +44,7 @@ include(source/ANALYSIS/XLMS/sources.cmake)
 include(source/PROCESSING/BASELINE/sources.cmake)
 include(source/PROCESSING/FILTERING/sources.cmake)
 include(source/PROCESSING/RESAMPLING/sources.cmake)
-include(source/PROCESSING/DATAREDUCTION/sources.cmake)
+include(source/PROCESSING/MISC/sources.cmake)
 include(source/PROCESSING/DEISOTOPING/sources.cmake)
 include(source/PROCESSING/CALIBRATION/sources.cmake)
 include(source/PROCESSING/SMOOTHING/sources.cmake)
@@ -112,7 +112,7 @@ include(include/OpenMS/COMPARISON/sources.cmake)
 
 include(include/OpenMS/PROCESSING/BASELINE/sources.cmake)
 include(include/OpenMS/PROCESSING/CALIBRATION/sources.cmake)
-include(include/OpenMS/PROCESSING/DATAREDUCTION/sources.cmake)
+include(include/OpenMS/PROCESSING/MISC/sources.cmake)
 include(include/OpenMS/PROCESSING/DEISOTOPING/sources.cmake)
 include(include/OpenMS/PROCESSING/ID/sources.cmake)
 include(include/OpenMS/PROCESSING/NOISEESTIMATION/sources.cmake)
