+++ src/openms/source/PROCESSING/FEATURE/sources.cmake
@@ -1,12 +1,9 @@
 ### the directory name
-set(directory source/PROCESSING/DATAREDUCTION)
+set(directory source/PROCESSING/FEATURE)
 
 ### list all filenames of the directory here
 set(sources_list
-DataFilters.cpp
 FeatureOverlapFilter.cpp
-SplineInterpolatedPeaks.cpp
-SplinePackage.cpp
 )
 
 ### add path to the filenames
@@ -19,5 +16,5 @@ endforeach(i)
 set(OpenMS_sources ${OpenMS_sources} ${sources})
 
 ### source group definition
-source_group("Source Files\\FILTERING\\DATAREDUCTION" FILES ${sources})
+source_group("Source Files\\FILTERING\\FEATURE" FILES ${sources})
