+++ src/openms/source/PROCESSING/MISC/DataFilters.cpp
@@ -6,7 +6,7 @@
 // $Authors: Marc Sturm $
 // --------------------------------------------------------------------------
 
-#include <OpenMS/PROCESSING/DATAREDUCTION/DataFilters.h>
+#include <OpenMS/PROCESSING/MISC/DataFilters.h>
 #include <OpenMS/KERNEL/Feature.h>
 #include <OpenMS/METADATA/MetaInfo.h>
 #include <OpenMS/KERNEL/ConsensusFeature.h>
