+++ src/openms/source/PROCESSING/MISC/SplineInterpolatedPeaks.cpp
@@ -7,8 +7,8 @@
 // --------------------------------------------------------------------------
 
 #include <OpenMS/KERNEL/StandardTypes.h>
-#include <OpenMS/PROCESSING/DATAREDUCTION/SplinePackage.h>
-#include <OpenMS/PROCESSING/DATAREDUCTION/SplineInterpolatedPeaks.h>
+#include <OpenMS/PROCESSING/MISC/SplinePackage.h>
+#include <OpenMS/PROCESSING/MISC/SplineInterpolatedPeaks.h>
 #include <OpenMS/KERNEL/MSSpectrum.h>
 #include <OpenMS/KERNEL/MSExperiment.h>
