+++ src/openms_gui/include/OpenMS/VISUAL/DIALOGS/DataFilterDialog.h
@@ -12,7 +12,7 @@
 // OpenMS_GUI config
 #include <OpenMS/VISUAL/OpenMS_GUIConfig.h>
 
-#include <OpenMS/PROCESSING/DATAREDUCTION/DataFilters.h>
+#include <OpenMS/PROCESSING/MISC/DataFilters.h>
 
 #include <QDialog>
