+++ src/openms_gui/include/OpenMS/VISUAL/VISITORS/LayerStoreData.h
@@ -10,7 +10,7 @@
 
 // OpenMS_GUI config
 
-#include <OpenMS/PROCESSING/DATAREDUCTION/DataFilters.h>
+#include <OpenMS/PROCESSING/MISC/DataFilters.h>
 #include <OpenMS/KERNEL/ConsensusMap.h>
 #include <OpenMS/KERNEL/FeatureMap.h>
 #include <OpenMS/KERNEL/MSExperiment.h>
